# fix(client): Implement readBytes in NetworkClient for faster downloads