Fixes: https://github.com/espressif/arduino-esp32/issues/9822

## Test downloading 22MB zip and using `readBytes` with ESP32-S3

Test results are relative and actual speeds will vary based on the environment. 

### Download from HTTP server (no SSL):
```cpp
// python3 -m http.server 8080 -d .
const char *fileDownloadUrl = "http://192.168.254.91:8080/esp32-3.0.1.zip";
```

New Method:
```
Download completed!
Total download time: 40.714000 s
Average download speed: 573.149578 Kb/s
```

Old Method:
```
Download completed!
Total download time: 63.890000 s
Average download speed: 365.240443 Kb/s
```

### Download from HTTPS server (with SSL):
```cpp
const char *fileDownloadUrl = "https://github.com/espressif/arduino-esp32/releases/download/3.0.1/esp32-3.0.1.zip";
```

New Method:
```
Download completed!
Total download time: 80.371000 s
Average download speed: 290.343680 Kb/s
```

Old Method:
```
Download completed!
Total download time: 162.368000 s
Average download speed: 143.718047 Kb/s
```