@JAndrassy 2024-06-11
readBytes is a function fo Stream class and should wait Stream's timeout for next byte
https://www.arduino.cc/reference/en/language/functions/communication/stream/streamreadbytes/