@me-no-dev 2024-06-11
@JAndrassy then maybe different solution is needed. reading network streams byte by byte is never a good idea