@me-no-dev 2024-06-11
@JAndrassy I got it, but readBytes reads byte by byte and that is also not good for Network, is what I meant