@me-no-dev 2024-06-11
So you suggest to leave Client to read byte by byte and waste a lot of time doing it?