@JAndrassy 2024-06-11
> So you suggest to leave Client to read byte by byte and waste a lot of time doing it?

there is `read)buffer, size)`. that should be used.