@me-no-dev 2024-06-11
`Stream.readBytes` says nowhere that we should be waiting for every single byte, therefore we can wait when there are no bytes. Do you get this, or are we just arguing for no reason?