@SuGlider 2024-06-11
About `Stream::readBytes()` --> it uses `Stream::timedRead()`. Setting timeout to Zero using `Stream::setTimeout(0)` may help WiFi streams using such function.

This PR should created some other Function Name... Or at least, test timeout for Zero and in case it is not zero, use the Stream Class original function. 

This PR implementation is equivalent to ignoring Timeout setup.