@SuGlider 2024-06-11
Please check https://github.com/espressif/arduino-esp32/issues/9823#issuecomment-2160705738
This is why timeout is important for certain Network operations.