@me-no-dev 2024-06-12
@SuGlider yes. It should wait the timeout, but should not read a byte by byte with timeout. Nobody said this is final implementation. 

This is network client, it gets many bytes. Currently the code that triggered this was downloading a large file from internet. Now imagine reading this file byte by byte and every time checking if more are available, if it got that single byte before that timeout and so on. Result is that you can not handle anything with speeds above 1Mbps on an S3 (I imagine much less on single core and slow RV chips). If we instead grab all bytes that are available at once and wait up to the timeout for more, then we can achieve speeds 3 times higher. I hope this makes it a bit more clear for you @SuGlider and @JAndrassy