@hitecSmartHome 2024-06-12
I'm more than happy to test if it will be compatible with arduino 2.0.17