@me-no-dev 2024-06-12
@hitecSmartHome you can try the current changes on top of WiFiClient