@JAndrassy 2024-06-12
so the now updated implementation in this PR should use `getTimeout()`, not `_timeout;`.

this implementation of readBytes will be used only if invoked on pointer of reference of type NetworkClient. if the pointer is  ESPLwIPClient, Client or Stream type then the reaBytes from Stream will be executed.