@me-no-dev 2024-06-12
>  if the pointer is ESPLwIPClient, Client or Stream type then the reaBytes from Stream will be executed.

That is correct. Stream implementation works fine. This is to fix it for NetworkClient and NetworkClientSecure (and HTTPClient)